% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calculate_residual_conductance.R
\name{calculate_rate_of_change}
\alias{calculate_rate_of_change}
\title{calculate_rate_of_change}
\usage{
calculate_rate_of_change(droughtbox_data)
}
\arguments{
\item{droughtbox_data}{}
}
\value{
A dataframe containing the strain_number, set temperature and the
rate of change between weight and time.
}
\description{
This function calculates the rate of change (aka slope) of the weight of a
small branch measured in grams and time measured in seconds.
}
\examples{
path_to_droughtbox_data <- system.file("extdata",
                            "acacia_aneura_25c.dat",
                            package = "HIEdroughtbox")

droughtbox_data <- read_hie_droughtbox_data_file(path_to_droughtbox_data)

calculate_rate_of_change(droughtbox_data = droughtbox_data)

}
