% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_calculate_residial_temperature_dependece.R
\name{calculate_residual_temperature_dependence}
\alias{calculate_residual_temperature_dependence}
\title{calculate_residual_temperature_dependence}
\usage{
calculate_residual_temperature_dependence(gmin, temperature)
}
\arguments{
\item{gmin}{Column containing the residual conductance values. The units
should be micro-mol*cm-2*s-1. The necessary data can be obtained with the
`calculate_residual_conductance` function.}

\item{temperature}{Column containing the temperature at which gmin was
measured values. The units should be degrees Celsius. The necessary data can
be obtained with the `calculate_residual_conductance` function.}
}
\value{
A dataframe with different parameters. Use the function
`plot_arrhenius` to select the best set of parameters according to Billon
et al 2020.
}
\description{
This function calculates a slope and a intercept with a varying number of
data points in a range from two to eight. For example it calculates a slope
using the first two data points (slope_1, intercept_1) and then uses the rest
for calculating a second slope and intercept (slope_1, intercept_1).
}
\examples{
path_to_tp_data <- system.file("extdata",
                            "tp_data.csv",
                             package = "HIEdroughtbox")

tp_data <- read.csv(path_to_tp_data, header = TRUE)

calculate_residual_temperature_dependence(gmin = tp_data$g_min,
                                        temperature = tp_data$temperature)

}
