% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/05_test-calculate_residual_temperature_dependence.R
\name{calculate_residual_temperature_dependence_purrr}
\alias{calculate_residual_temperature_dependence_purrr}
\title{calculate_residual_temperature_dependence_purrr}
\usage{
calculate_residual_temperature_dependence_purrr(
  data,
  number_of_temperature_steps
)
}
\arguments{
\item{data}{Dataframe with two columns, temperature and residual conductance}

\item{number_of_temperature_steps}{Integer (between 6 and 8) indicating the total number of temperatures at
which gmin was measured.}
}
\value{
A dataframe with different parameters. Use the function
`plot_arrhenius` to select the best set of parameters according to Billon
et al 2020.
}
\description{
This function is a reimplementation of `calculate_residual_temperature_dependence`
but with purrr for avoiding the for loop. It should give the same result.


Calculates a slope and a intercept with a varying number of
data points in a range from two to eight. For example it calculates a slope
using the first two data points (slope_regression_1, intercept_regression_1)
and then uses the rest for calculating a second slope and intercept
(slope_regression_2, intercept_regression_2).
}
\examples{
\dontrun{
path_to_tp_data <- system.file("extdata",
                            "tp_data.csv",
                             package = "HIEdroughtbox")

tp_data <- read.csv(path_to_tp_data, header = TRUE)

calculate_residual_temperature_dependence_purrr(tp_data,
                                               number_of_temperature_steps = 8)
}
}
