% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calculate_residual_conductance.R
\name{calculate_transpiration_rates}
\alias{calculate_transpiration_rates}
\title{calculate_transpiration_rates}
\usage{
calculate_transpiration_rates(droughtbox_data, leaf_and_branch_area_data)
}
\arguments{
\item{droughtbox_data}{Dataframe loaded with the function
`read_hie_droughtbox_data()`.

This dataframe must contain the following columns:
 strain_avg_1_microstrain_avg
 strain_avg_2_microstrain_avg
 strain_avg_3_microstrain_avg
 strain_avg_4_microstrain_avg

 set_point_t_avg_avg
 vpd_avg_kpa_avg
 date_time}

\item{leaf_and_branch_area_data}{Dataframe containing the leaf and/or branch
areas.

This dataframe must contain the following columns:
 areas_cm2
 strain_number
 set_temperature
 tree_id}
}
\value{
A dataframe
}
\description{
This function calculates the transpiration rate of a small branch following
equation 1 found the research paper titled 'The DroughtBox: A new tool for
phenotyping residual branch conductance and its temperature dependence during
drought' by Billon and colleagues.
}
\examples{
\dontrun{path_droughtbox_leaf_branch_areas <- system.file("extdata",
                                                "acacia_aneura_leaf_branch_areas.xlsx",
                                                package = "HIEdroughtbox")

path_to_droughtbox_data <- system.file("extdata",
                            "acacia_aneura_25c.dat",
                            package = "HIEdroughtbox")

droughtbox_data <- read_hie_droughtbox_data_file(path_to_droughtbox_data)
species_areas <- readxl::read_excel(path_droughtbox_leaf_branch_areas)

calculate_residual_conductance(droughtbox_data = droughtbox_data,
                               leaf_and_branch_area_data = species_areas)}

}
