% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_droughtbox_utils.R
\name{clean_droughtbox_colnames}
\alias{clean_droughtbox_colnames}
\title{Title}
\usage{
clean_droughtbox_colnames(path_data_droughtbox)
}
\arguments{
\item{path_data_droughtbox}{}
}
\description{
Title
}
