% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_droughtbox_utils.R
\name{clean_droughtbox_data}
\alias{clean_droughtbox_data}
\title{clean_droughtbox_data}
\usage{
clean_droughtbox_data(droughtbox_data, remove_n_observations, threshold = 0.2)
}
\arguments{
\item{droughtbox_data}{Dataframe loaded with the function.
`read_hie_droughtbox_data`}

\item{remove_n_observations}{Integer indicating the number of values that
need to be removed at the beginning and at the end of each tare_count group.}

\item{threshold}{Float indicating the threshold at which values should be
removed}
}
\value{
A dataset
}
\description{
This function removes wrong data points that are produced by the Droughtbox
after each taring process. First it removes values lower than a `threshold`,
(which is in grams) and then it removes the first and last values of each
tare_count

The function `plot_raw_strains_weights` can be used to visualize the data
points might need to be removed.
}
\examples{
data <- read_hie_droughtbox_data("inst/extdata/acacia_aneura_25c.dat")

# Remove tare_counts without enough measurements
data <- data \%>\% filter(!tare_count_smp \%in\% c("13","14","28"))

# Clean data
clean_droughtbox_data(data, remove_n_observations = 6)
}
