% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_droughtbox_utils.R
\name{create_empty_droughtbox_leaf_branch_areas_sheet}
\alias{create_empty_droughtbox_leaf_branch_areas_sheet}
\title{create_empty_droughtbox_leaf_branch_areas_sheet}
\usage{
create_empty_droughtbox_leaf_branch_areas_sheet(save_empty_df_at = NULL)
}
\arguments{
\item{save_empty_df_at}{String indicating the path where the empty datasheet
for recording the leaf and branch areas should be saved.}
}
\value{
Empty dataframe filled with NAs.
}
\description{
This function generates a CSV files filled with NA'S with the required
columns for recording the leaf and branch area data necessary to estimate
gmin or gres.
}
\examples{
\dontrun{create_empty_droughtbox_leaf_branch_areas_sheet("path/to/folder")}

}
