% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_droughtbox_utils.R
\name{filter_droughtbox_data}
\alias{filter_droughtbox_data}
\title{read_hie_droughtbox_leaf_branch_areas}
\usage{
filter_droughtbox_data(
  droughtbox_data,
  from_start_date = NULL,
  to_end_date = NULL,
  from_start_time = NULL,
  to_end_time = NULL
)
}
\arguments{
\item{droughtbox_data}{Dataframe loaded with the function
`read_hie_droughtbox_data`.}

\item{from_start_date}{String indicating the initial Year, Month and Day to
filter in the dataset. It must have a YYYY-MM-DD format.}

\item{to_end_date}{String indicating the final Year, Month and Day to filter
in the dataset. It must have a YYYY-MM-DD format.}

\item{from_start_time}{String indicating the initial hour, minutes and
seconds to filter in the dataset. It must have a HH:MM:SS format.}

\item{to_end_time}{String indicating the final hour, minutes and seconds
to filter in the dataset. It must have a HH:MM:SS format.}
}
\value{
Dataframe with the selected dates and times.
}
\description{
This function is meant to be used to removed chunks of data that is collected
when the droughtbox has not reach the climatic conditions desired.

This functions does not remove individual observations.
}
\details{
description
This function reads CSV files containing information about the leaf and/or
branch area of the samples measured in the droughtbox.

To create a datasheet with the required information, run the
function `create_empty_droughtbox_leaf_branch_areas_sheet('path/to/folder')`.
with the path to the folder where the datasheet should be saved.

The CSV file MUST contain the following columns:

set_temperature: Integer indicating the temperature at which gmin/gres was
measured.

strain_number: Integer indicating in which of the four hooks the the sample
was positioned.

leaf_area: Float in cm2 with the total leaf area of the sample.

The following columns are optional:

tree_id: String with an unique code identifying each sample.

surface_branch_area: Float in cm2 with the area of the branch without any
leaves attached. If not provided, the total surface_branch_area can be
approximated using the formula A = pi*radius*(length + radius), assuming that
the branch has a cone shape.

branch_basal_diameter_mm: Float in millimeters indicating the basal diameter
of the sample.

branch_length_cm: Float in centimeters indicating the total length of the
sample.

param path_droughtbox_leaf_branch_areas String indicating the location of
the CSV file in your computer.

return A dataframe.

importFrom magrittr %>%

examples
path_droughtbox_leaf_branch_areas <- system.file("extdata",
                                                "acacia_aneura_leaf_branch_areas.xlsx",
                                                package = "HIEdroughtbox")

read_hie_droughtbox_leaf_branch_areas(path_droughtbox_leaf_branch_areas)

export
filter_droughtbox_data
}
\examples{
\dontrun{create_empty_droughtbox_leaf_branch_areas_sheet("path/to/folder")}#' @importFrom magrittr \%>\%

path_to_droughtbox_data <- system.file("extdata",
                            "acacia_aneura_25c.dat",
                            package = "HIEdroughtbox")

droughtbox_data <- read_hie_droughtbox_data_file(path_to_droughtbox_data)

filter_droughtbox_data(droughtbox_data,
                            from_start_date = "2024/03/04",
                            to_end_date = "2024/03/04",
                            from_start_time = NULL,
                            to_end_time = NULL)

filter_droughtbox_data(droughtbox_data,
                            from_start_date = "2024/03/04",
                            to_end_date = "2024/03/04",
                            from_start_time = "12:51:00",
                            to_end_time = "12:52:00")

filter_droughtbox_data(droughtbox_data,
                            from_start_date = NULL,
                            to_end_date = NULL,
                            from_start_time = "12:51:00",
                            to_end_time = "12:52:00")

}
