% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_droughtbox_utils.R
\name{merge_droughtbox_data}
\alias{merge_droughtbox_data}
\title{merge_droughtbox_data}
\usage{
merge_droughtbox_data()
}
\arguments{
\item{droughtbox_data}{Dataframe loaded with the function.
`read_hie_droughtbox_data`}
}
\value{
A dataset
}
\description{
This function
}
