% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_droughtbox_plots.R
\name{plot_arrhenius}
\alias{plot_arrhenius}
\title{plot_arrhenius}
\usage{
plot_arrhenius(residual_conductance_data)
}
\arguments{
\item{residual_conductance_data}{Dataframe obtained with the
`calculate_residual_conductance` function}
}
\value{
A ggplot2 object with the weight (grams) measured by each strain (4 in
total) inside the Droughtbox.
}
\description{
plot_arrhenius
}
