% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_droughtbox_plots.R
\name{plot_droughtbox_climatic_controls}
\alias{plot_droughtbox_climatic_controls}
\title{plot_droughtbox_climatic_controls}
\usage{
plot_droughtbox_climatic_controls(droughtbox_data, cowplot = TRUE)
}
\arguments{
\item{droughtbox_data}{Dataframe loaded with the function
`read_hie_droughtbox_data`}

\item{cowplot}{Boolean indicating if plots should be arranged in one single
figure with 2 columns and 2 rows (TRUE) or if should be arranged individually}
}
\value{
A ggplot2 object with 4 figures
}
\description{
plot_droughtbox_climatic_controls
}
