% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_droughtbox_plots.R
\name{plot_raw_strains_weights}
\alias{plot_raw_strains_weights}
\title{plot_raw_strains_weights}
\usage{
plot_raw_strains_weights(droughtbox_data)
}
\arguments{
\item{droughtbox_data}{Dataframe loaded with the function
`read_hie_droughtbox_data`}
}
\value{
A ggplot2 object with the weight (grams) measured by each strain (4 in
total) inside the Droughtbox
}
\description{
This function displays the raw weights (grams) measured inside the Droughtbox
}
\examples{
droughtbox_data <- read_hie_droughtbox_data("data/acacia_aneura_25c.dat")
plot_raw_strains_weights(droughtbox_data)
}
