% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_droughtbox_utils.R
\name{read_hie_droughtbox_data}
\alias{read_hie_droughtbox_data}
\title{read_hie_droughtbox_data}
\usage{
read_hie_droughtbox_data(path_data_droughtbox)
}
\arguments{
\item{path_data_droughtbox}{String indicating the location of the .dat file in your computer}
}
\value{
A dataframe with 25 columns
}
\description{
This function reads the raw .dat file downloaded from the droughtbox located
at the Hawkesbury Institute for the Environment
}
\examples{
read_hie_droughtbox_data(path/to/file.dat)


}
