% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_droughtbox_utils.R
\name{read_hie_droughtbox_data_file}
\alias{read_hie_droughtbox_data_file}
\title{read_hie_droughtbox_data_file}
\usage{
read_hie_droughtbox_data_file(path_droughtbox_data_file)
}
\arguments{
\item{path_droughtbox_data_file}{String indicating the location of the .dat
file in your computer.}
}
\value{
A dataframe with 25 columns.
}
\description{
This function reads the raw .dat file downloaded from the droughtbox located
at the Hawkesbury Institute for the Environment.
}
\examples{
path_to_droughtbox_data_file <- system.file("extdata",
                                            "acacia_aneura_25c.dat",
                                            package = "HIEdroughtbox")

read_hie_droughtbox_data_file(path_to_droughtbox_data_file)

}
