% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_droughtbox_utils.R
\name{read_hie_droughtbox_data_folder}
\alias{read_hie_droughtbox_data_folder}
\title{read_hie_droughtbox_data_folder}
\usage{
read_hie_droughtbox_data_folder(path_droughtbox_data_folder)
}
\arguments{
\item{path_droughtbox_data_folder}{String indicating the location of a folder
containing 2 or more .dat files in your computer.}
}
\value{
A single list with 2 or more data frames.
}
\description{
This function reads 2 or more raw .dat files downloaded from the
droughtbox located at the Hawkesbury Institute for the Environment.
}
\examples{
path_to_droughtbox_data_folder <- system.file("extdata",
                                               package = "HIEdroughtbox")

read_hie_droughtbox_data_folder(path_to_droughtbox_data_folder)

}
