% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_calculate_relative_water_content.R
\name{relative_water_content}
\alias{relative_water_content}
\title{Relative water content}
\usage{
relative_water_content(droughtbox_data, dry_weights)
}
\arguments{
\item{droughtbox_data}{Dataframe loaded with the function
`read_hie_droughtbox_data()`.}

\item{dry_weights}{Dataframe loaded with the functionv `read_dry_weights_data()`}
}
\description{
This function calculates the relative water content of each sample.
}
