% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_droughtbox_utils.R
\name{reshape_droughtbox_data}
\alias{reshape_droughtbox_data}
\title{reshape_droughtbox_data}
\usage{
reshape_droughtbox_data(droughtbox_data)
}
\arguments{
\item{droughtbox_data}{}
}
\value{
A dataframe
}
\description{
This function transform the droughtbox data from wide to a long format
}
