% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calculate_minimum_conductance.R
\name{residual_conductance}
\alias{residual_conductance}
\title{residual_conductance}
\usage{
residual_conductance(
  droughtbox_data,
  leaf_and_stem_area,
  calculate_gres = TRUE
)
}
\arguments{
\item{leaf_and_stem_area}{Dataframe loaded with the function
`read_hie_droughtbox_leaf_branch_areas`


containing leaf area and/or branch
surface area. This}

\item{calculate_gres}{Boolean indicating if residual branch conductance (TRUE)
or leaf residual conductance (FALSE) should be calculated}
}
\description{
This function calculates residual leaf conductance (gmin) or residual
branch conductance (gres) following equation 2 found the research paper
titled 'The DroughtBox: A new tool for phenotyping residual branch
conductance and its temperature dependence during drought' by Billon and
colleges.

s @param droughtbox_data Dataframe loaded with the function
`read_hie_droughtbox_data()`
}
