% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calculate_residual_conductance.R
\name{calculate_residual_conductance}
\alias{calculate_residual_conductance}
\title{calculate_residual_conductance}
\usage{
calculate_residual_conductance(droughtbox_data, leaf_and_branch_area_data)
}
\arguments{
\item{leaf_and_branch_area_data}{Dataframe loaded with the function
`read_hie_droughtbox_leaf_branch_areas()`.}
}
\value{
A dataframe with the species_name, sample_id, strain_number,
set_temperature, transpiration_grams_per_sec_cm2 and
median_vpd residual_conductance as columns.
}
\description{
This function calculates residual leaf conductance (gmin) or residual
branch conductance (gres) following equation 2 found the research paper
titled 'The DroughtBox: A new tool for phenotyping residual branch
conductance and its temperature dependence during drought' by Billon and
colleagues.

s @param droughtbox_data Dataframe loaded with the function
`read_hie_droughtbox_data()`.
}
\examples{
path_droughtbox_leaf_branch_areas <- system.file("extdata",
                                                "acacia_aneura_leaf_branch_areas.csv",
                                                package = "HIEdroughtbox")

path_to_droughtbox_data <- system.file("extdata",
                            "acacia_aneura_25c.dat",
                            package = "HIEdroughtbox")

droughtbox_data <- read_hie_droughtbox_data(path_to_droughtbox_data)
species_areas <- read_hie_droughtbox_leaf_branch_areas(path_droughtbox_leaf_branch_areas)

calculate_residual_conductance(droughtbox_data = droughtbox_data,
                               leaf_and_branch_area_data = species_areas)

}
