% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_droughtbox_utils.R
\name{clean_droughtbox_colnames}
\alias{clean_droughtbox_colnames}
\title{clean_droughtbox_colnames}
\usage{
clean_droughtbox_colnames(path_data_droughtbox)
}
\arguments{
\item{path_data_droughtbox}{String indicating the location of the .dat file in your computer}
}
\value{
Vector of strings of length 30
}
\description{
This is an internal function meant to be used inside the
`clean_droughtbox_data` function.

First it merges the the first two rows (which contain the units and the data
type of each column) of the .dat file. and then merges those merged rows with
each colname.

The .dat file is downloaded from the the droughtbox

The pattern of the new colname is varname_unit_data_type. For example
"air_tc_avg_deg_c_avg" the varname is air_tc_avg, the unit is deg_c and the
data_type is avg.

Some colnames don`t have a units or data type. For example tare_count_sm,
where the varname is tare_count and the data_type is sm.
}
\examples{
clean_droughtbox_colnames(path/to/file.dat)

}
