% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_droughtbox_utils.R
\name{clean_droughtbox_data}
\alias{clean_droughtbox_data}
\title{Title}
\usage{
clean_droughtbox_data(path_data_droughtbox)
}
\arguments{
\item{path_data_droughtbox}{String indicating the location of the .dat file in your computer}
}
\description{
Title
}
