% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_droughtbox_utils.R
\name{filter_droughtbox_data}
\alias{filter_droughtbox_data}
\title{filter_droughtbox_data}
\usage{
filter_droughtbox_data(
  droughtbox_data,
  from_start_date = NULL,
  to_end_date = NULL,
  from_start_time = NULL,
  to_end_time = NULL
)
}
\arguments{
\item{droughtbox_data}{Dataframe loaded with the function
`read_hie_droughtbox_data`}

\item{from_start_date}{String indicating the initial Year, Month and Day to
filter in the dataset. It must have a YYYY-MM-DD format}

\item{to_end_date}{String indicating the final Year, Month and Day to filter
in the dataset. It must have a YYYY-MM-DD format}

\item{from_start_time}{String indicating the initial hour, minutes and seconds
to filter in the dataset. It must have a HH:MM:SS format}

\item{to_end_time}{String indicating the final hour, minutes and seconds
to filter in the dataset. It must have a HH:MM:SS format}
}
\value{
Dataframe with the selected dates and times
}
\description{
This function is meant to be used to removed chunks of data that is collected
when the droughtbox has not reach the climatic conditions desired.

This functions does not remove individual observations
}
\examples{
path_to_droughtbox_data <- system.file("extdata",
                            "acacia_aneura_25c.dat",
                            package = "HIEdroughtbox")

droughtbox_data <- read_hie_droughtbox_data(path_to_droughtbox_data)

filter_droughtbox_data(droughtbox_data,
                            from_start_date = "2024/03/04",
                            to_end_date = "2024/03/04",
                            from_start_time = NULL,
                            to_end_time = NULL)

filter_droughtbox_data(droughtbox_data,
                            from_start_date = "2024/03/04",
                            to_end_date = "2024/03/04",
                            from_start_time = "12:51:00",
                            to_end_time = "12:52:00")

filter_droughtbox_data(droughtbox_data,
                            from_start_date = NULL,
                            to_end_date = NULL,
                            from_start_time = "12:51:00",
                            to_end_time = "12:52:00")
}
