% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_droughtbox_utils.R
\name{merge_droughtbox_data}
\alias{merge_droughtbox_data}
\title{merge_droughtbox_data}
\usage{
merge_droughtbox_data(...)
}
\arguments{
\item{...}{n number of objects of class data.frame.}
}
\value{
A single data.frame object.
}
\description{
This function is meant to be used as the last step previously to calculate
the residual conductace. It takes any number of prevously cleaned .dat files
and merge them into a single data frame.
}
\examples{
\dontrun{merge_droughtbox_data(clean_data_1, clean_data_2, clean_data_n)}

}
