% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_droughtbox_plots.R
\name{plot_droughtbox_climatic_controls}
\alias{plot_droughtbox_climatic_controls}
\title{plot_droughtbox_climatic_controls}
\usage{
plot_droughtbox_climatic_controls(droughtbox_data, cowplot = TRUE)
}
\arguments{
\item{droughtbox_data}{Dataframe loaded with the function
`read_hie_droughtbox_data`}

\item{cowplot}{Boolean indicating if plots should be arranged in one single
figure with 2 columns and 3 rows (TRUE) or if should be arranged individually}
}
\value{
A ggplot2 object with a total of 5 figures
}
\description{
This function plots the change across time of the climatic controls specified
by the user (shown in red) and the climatic controls measured inside of the
Droughtbox
}
\examples{
droughtbox_data <- read_hie_droughtbox_data("inst/extdata/acacia_aneura_25c.dat")
plot_droughtbox_climatic_controls(droughtbox_data, cowplot = TRUE)
}
